[MIG] runbot_skip_test -> 9.0
