+++ runbot_skip_tests/models/__init__.py
@@ -3,4 +3,6 @@
 #    (<http://www.savoirfairelinux.com>).
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import models
+from . import (
+    runbot_build,
+)
